+++ tests/test/org/commcare/test/utilities/TemplateStructureTest.java
@@ -1,9 +1,8 @@
-package org.commcare.test.utils.test;
+package org.commcare.test.utilities;
 
 import org.commcare.util.mocks.SessionWrapper;
 import org.junit.Assert;
 
-import org.commcare.test.utils.MockApp;
 import org.commcare.util.SessionFrame;
 import org.junit.Before;
 import org.junit.Test;
