+++ tests/test/org/commcare/test/utilities/TestHelpers.java
@@ -1,4 +1,4 @@
-package org.commcare.test.utils;
+package org.commcare.test.utilities;
 
 import org.junit.Assert;
